Generate the cocoapod from the root project
