+++ src/commonMain/kotlin/HelloWorld.kt
@@ -1,0 +1,3 @@
+object HelloWorld {
+    val message = "Hello World!"
+}
